# apps/pairing-server/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DropVoice Pairing & Signaling Server API",
    "description": "公网 HTTPS 配对 + WebRTC 信令服务器。部署在 VPS 上，经 Cloudflare 回源到 Caddy:4443，再反向代理到 axum 127.0.0.1:38424。\n\n## 职责边界\n本服务只做**设备注册 + WebRTC 信令转发**，绝不中继文本数据。\n手机扫码后发起 offer，服务器通过 SSE 转发给桌面，桌面回填 answer；\nSDP 交换完成后，手机↔桌面 WebRTC DataChannel P2P 直连，服务器退出。\n\n## 认证\n- `POST /api/devices`：新建设备无需认证；已存在设备须携带 Bearer（当前 token）方可复用/轮换，否则 401——token 绝不复述给无凭据方。\n- `PUT /api/devices/{id}/status`：Bearer token（注册时颁发的 `pairing_token`）。\n- `POST /api/devices/{id}/webrtc/offer`：body 带 code（首次配对）或 token（重连）。\n- `POST /api/devices/{id}/webrtc/subscribe`：Bearer token → 一次性 SSE 票据。\n- `GET /api/devices/{id}/webrtc/events`（SSE）：query `?ticket=`（一次性票据）。\n- `POST /api/devices/{id}/webrtc/answer`：Bearer token。\n- `GET /api/devices/{id}/webrtc/answer/{session_id}`：session_id 自身即为凭据。",
    "contact": {
      "name": "DropVoice"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "https://ps.dropvoice.online",
      "description": "生产环境（Cloudflare → Caddy:4443 → axum:38424）"
    }
  ],
  "paths": {
    "/api/devices": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "设备注册（凭据受控的幂等 upsert）",
        "description": "桌面端本地生成持久化 UUID v4 作为 device_id（首次启动后永不变更），\ntoken 持久化在桌面本地；重复注册需携带当前有效 token（Authorization Bearer）：\n- 新 device_id → 创建，颁发 pairing_token → 201\n- 已存在 + 有效 Bearer → 200（token 回显；超过 24h TTL 则轮换为新 token）\n- 已存在 + 无/无效 Bearer → 401（token 绝不复述给无凭据方）",
        "operationId": "registerDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegisterRequest"
              },
              "examples": {
                "new_device": {
                  "summary": "首次注册",
                  "value": {
                    "address": {
                      "ip": "192.168.1.100",
                      "port": 38425
                    },
                    "device_id": "550e8400-e29b-41d4-a716-446655440000",
                    "device_name": "My PC",
                    "platform": "desktop"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "设备已存在（Bearer 校验通过）。token 为回显或轮换后的新值",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRegisterResponse"
                }
              }
            }
          },
          "201": {
            "description": "新建设备成功，颁发 token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRegisterResponse"
                },
                "example": {
                  "address": {
                    "ip": "192.168.1.100",
                    "port": 38425
                  },
                  "created_at": "2026-07-18T12:00:00Z",
                  "device_name": "My PC",
                  "id": "550e8400-e29b-41d4-a716-446655440000",
                  "pairing_token": "a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2",
                  "platform": "desktop"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/devices/{device_id}/status": {
      "put": {
        "tags": [
          "devices"
        ],
        "summary": "状态/地址上报（心跳）",
        "description": "桌面端周期性上报（DEVICE_STATUS_INTERVAL 5 分钟），IP 变化时立即触发额外上报\n（节流 IP_REPORT_THROTTLE 10 秒）。用于：① 续期 token；② 标记设备在线；\n③ IP 变化时刷新服务端记录。\n返回 401 表示 token 过期，客户端应自动重 POST /devices（同 device_id）触发续期，\n再重放本次心跳（§7.5）。",
        "operationId": "reportStatus",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "description": "客户端生成的 UUID v4 设备标识",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusReportRequest"
              },
              "examples": {
                "heartbeat": {
                  "summary": "周期心跳",
                  "value": {
                    "address": {
                      "ip": "192.168.1.100",
                      "port": 38425
                    }
                  }
                },
                "ip_changed": {
                  "summary": "IP 变化上报",
                  "value": {
                    "address": {
                      "ip": "192.168.1.105",
                      "port": 38425
                    },
                    "device_name": "Renamed PC"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "上报成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusReportResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceAddress": {
        "type": "object",
        "description": "设备局域网地址（B 线桌面端本地服务器地址）。",
        "required": [
          "ip",
          "port"
        ],
        "properties": {
          "ip": {
            "type": "string",
            "description": "设备局域网 IPv4/IPv6 地址。",
            "example": "192.168.1.100"
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "description": "桌面端本地服务器端口（默认 38425）。",
            "example": 38425,
            "minimum": 0
          }
        }
      },
      "DeviceRegisterRequest": {
        "type": "object",
        "description": "`POST /devices` 请求体（与 `pairing-server.yaml` `DeviceRegisterRequest` 一致）。",
        "required": [
          "device_id",
          "platform",
          "address"
        ],
        "properties": {
          "device_id": {
            "type": "string",
            "format": "uuid",
            "description": "客户端生成的 UUID v4，持久化在 config.toml，永不变更。"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform",
            "description": "当前仅 desktop 会注册到配对服务器。"
          },
          "device_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "固化的设备名（§6.6）。可选字段；缺省时由服务端记录为空，\n手机端 UI 用 `name (ip)` 组合显示。",
            "maxLength": 64
          },
          "address": {
            "$ref": "#/components/schemas/DeviceAddress",
            "description": "桌面端当前局域网地址。"
          }
        }
      },
      "DeviceRegisterResponse": {
        "type": "object",
        "description": "`POST /devices` / 复用响应体（`DeviceRegisterResponse`）。",
        "required": [
          "id",
          "platform",
          "address",
          "pairing_token",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "等同请求的 `device_id`。"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "device_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "address": {
            "$ref": "#/components/schemas/DeviceAddress"
          },
          "pairing_token": {
            "type": "string",
            "description": "64 字符随机字符串，用于后续认证。续期时此字段更新。"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "错误响应体（与 `pairing-server.yaml#/components/schemas/Error` 一致）。",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "机器可读错误码，`SCREAMING_SNAKE_CASE`。",
            "example": "SESSION_NOT_FOUND"
          },
          "message": {
            "type": "string",
            "description": "人类可读错误信息。"
          },
          "details": {
            "type": "object",
            "description": "可选调试详情。"
          }
        }
      },
...
      "Platform": {
        "type": "string",
        "description": "设备平台。当前仅 desktop 会注册到配对服务器。",
        "enum": [
          "desktop"
        ],
        "example": "desktop"
      },
      "StatusReportRequest": {
        "type": "object",
        "description": "`PUT /devices/{id}/status` 请求体（`StatusReportRequest`）。",
        "required": [
          "address"
        ],
        "properties": {
          "address": {
            "$ref": "#/components/schemas/DeviceAddress"
          },
          "device_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "用户改名后携带，刷新服务端记录。",
            "maxLength": 64
          }
        }
      },
      "StatusReportResponse": {
        "type": "object",
        "description": "`PUT /devices/{id}/status` 响应体（`StatusReportResponse`）。",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数无效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            },
            "example": {
              "error": {
                "code": "INVALID_REQUEST",
                "details": {
                  "field": "device_id"
                },
                "message": "The request payload is invalid"
              }
            }
          }
        }
      },
...
      "NotFound": {
        "description": "设备不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            },
            "example": {
              "error": {
                "code": "DEVICE_NOT_FOUND",
                "message": "The device does not exist"
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "速率超限（RATE_LIMIT 1 req/IP/秒）",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            },
            "example": {
              "error": {
                "code": "RATE_LIMITED",
                "message": "Too many requests"
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Token 无效或已过期",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorBody"
            },
            "example": {
              "error": {
                "code": "TOKEN_INVALID",
                "message": "The pairing token is invalid or expired"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "注册时颁发的 64 字符 `pairing_token`。"
      }
    }
  },
...
}